[수정] 불필요한 코드를 제거하고, 인덱싱 수정
