+++ BlackCat/BaseArchitecture/BaseCollectionViewCell.swift
@@ -1,0 +1,22 @@
+//
+//  BaseUICollectionViewCell.swift
+//  BlackCat
+//
+//  Created by Hamlit Jason on 2022/10/19.
+//
+
+import UIKit
+
+class BaseCollectionViewCell: UICollectionViewCell {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        initialize()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func initialize() { /* override point */ }
+}
