+++ BlackCat/Filter/FilterCell/FilterBaseCell.swift
@@ -8,18 +8,6 @@
 import UIKit
 import RxSwift
 
-class FilterBaseCell: UICollectionViewCell {
+class FilterBaseCell: BaseCollectionViewCell {
     var disposeBag: DisposeBag = DisposeBag()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        
-        initialize()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func initialize() { /* override point */ }
 }
